Update group ID to xyz.voltawallet
