+++ module-core/build.gradle.kts
@@ -1,4 +1,4 @@
-import xyz.volta.Dependencies
+import xyz.voltawallet.Dependencies
 
 dependencies {
   implementation(Dependencies.Web3j.CORE)
