--- module-core/src/main/java/xyz/volta/VoltaClient.java
@@ -1,9 +1,0 @@
-package xyz.volta;
-
-import xyz.volta.exception.VoltaException;
-import xyz.volta.model.UserOperation;
-
-public interface VoltaClient {
-  UserOperation build(final UserOperation operation) throws VoltaException;
-  Object send(final UserOperation operation) throws VoltaException;
-}
