--- module-core/src/main/java/xyz/volta/exception/VoltaException.java
@@ -1,8 +1,0 @@
-package xyz.volta.exception;
-
-public final class VoltaException extends Exception {
-
-  public VoltaException(final String message) {
-    super(message);
-  }
-}
