--- module-core/src/main/java/xyz/volta/internal/model/VoltaError.java
@@ -1,14 +1,0 @@
-package xyz.volta.internal.model;
-
-public class VoltaError {
-  private int code;
-  private String message;
-
-  public int getCode() {
-    return code;
-  }
-
-  public String getMessage() {
-    return message;
-  }
-}
