+++ module-core/src/main/java/xyz/voltawallet/VoltaClient.java
@@ -1,0 +1,9 @@
+package xyz.voltawallet;
+
+import xyz.voltawallet.exception.VoltaException;
+import xyz.voltawallet.model.UserOperation;
+
+public interface VoltaClient {
+  UserOperation build(final UserOperation operation) throws VoltaException;
+  Object send(final UserOperation operation) throws VoltaException;
+}
