+++ module-core/src/main/java/xyz/voltawallet/exception/VoltaException.java
@@ -1,0 +1,8 @@
+package xyz.voltawallet.exception;
+
+public final class VoltaException extends Exception {
+
+  public VoltaException(final String message) {
+    super(message);
+  }
+}
