+++ module-core/src/main/java/xyz/voltawallet/internal/model/VoltaError.java
@@ -1,0 +1,14 @@
+package xyz.voltawallet.internal.model;
+
+public class VoltaError {
+  private int code;
+  private String message;
+
+  public int getCode() {
+    return code;
+  }
+
+  public String getMessage() {
+    return message;
+  }
+}
