+++ module-core/src/test/java/xyz/voltawallet/VoltaSdkTest.java
@@ -1,0 +1,18 @@
+package xyz.voltawallet;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class VoltaSdkTest {
+
+  @Test
+  public void nullUrl_MustThrow() {
+    assertThrows(IllegalArgumentException.class, () -> VoltaSdk.newVoltaClient(null));
+  }
+
+  @Test
+  public void emptyUrl_MustThrow() {
+    assertThrows(IllegalArgumentException.class, () -> VoltaSdk.newVoltaClient(null));
+  }
+}
