+++ module-example/build.gradle.kts
@@ -1,7 +1,7 @@
-import xyz.volta.Dependencies
+import xyz.voltawallet.Dependencies
 
 tasks.jar {
-  manifest.attributes["Main-Class"] = "xyz.volta.example.Main"
+  manifest.attributes["Main-Class"] = "xyz.voltawallet.example.Main"
   exclude("META-INF/BC1024KE.RSA", "META-INF/BC1024KE.SF", "META-INF/BC1024KE.DSA")
   exclude("META-INF/BC2048KE.RSA", "META-INF/BC2048KE.SF", "META-INF/BC2048KE.DSA")
   val dependencies = configurations
